+++ RELEASE_NOTES.md
@@ -12,6 +12,10 @@
 * Added a resend button to the Live Chat feature inside Custom Chatbox.
   Suggested by @notd3vil - *Add a "resend" button to messages sent in the chatbox*
 
+**World Modal**
+* Added the iOS World Size badge to world modal.
+Suggested by @octomiku01 - *Add IOS download size to world modal*
+
 **Fixes**
 
 * Fixed an issue where the KikitanXD mic input was reset after restarting.